backend/app/main.py
from fastapi import FastAPI
from dotenv import load_dotenv
from fastapi.middleware.cors import CORSMiddleware
from app.core.firebase_admin import init_firebase_admin
from app.auth.routes import router as auth_router

load_dotenv()

init_firebase_admin()

app = FastAPI()
# Language: Python
app.add_middleware(
    CORSMiddleware,
    allow_origins=[
        "http://localhost:3000",
        "http://127.0.0.1:3000",
    ],
    allow_credentials=True,
    allow_methods=["*"],
    allow_headers=["*"],
)

app.include_router(auth_router, prefix="/auth")

@app.get("/")
def health():
    return {"Status": "Backend running"}
