backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
python-dotenv==1.0.1

firebase-admin==6.5.0

SQLAlchemy==2.0.32
aiosqlite==0.20.0

passlib[bcrypt]==1.7.4
